Refactor task scheduler structure: rename packages and consolidate leader election interfaces
